+++ src/MyDraggableList/types.tsx
@@ -1,0 +1,9 @@
+export type Item = {
+  id: string;
+  content: string;
+};
+
+export type Props = {
+  items: Item[];
+  setData: (data: Item[]) => void;
+};
