+++ src/index.tsx
@@ -1,3 +1,3 @@
 import 'react-native-gesture-handler';
 
-export { MyButton } from './MyButton';
+export { MyDraggableList } from './MyDraggableList';
